Got it working
All four methods of changing the fuel bar segmentation now work properly.

Fun note: The fixed fuel segmentation stacks. So one could, say, make an outfit with "50", and then the player can pick what they want. Want each segment to be 250? Just install 5. Want it to be increments of 500? OK, install 10. Flying a Korath super tanker and want it to be increments of 5000? Sure, just install 100. Whatever you want.

+++ source/Ship.h
@@ -338,7 +338,7 @@ public:
 	bool HyperDriveFuelBar() const;
 	bool ScramDriveFuelBar() const;
 	bool JumpDriveFuelBar() const;
-	int FixedScaleFuelBar() const;
+	double FixedScaleFuelBar() const;
 	bool DisplayMass() const;
 	bool DisplayHyperFuelCost() const;
 	bool DisplayScramFuelCost() const;
